% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argsDefaults.R
\docType{data}
\name{chromToExclude_default}
\alias{chromToExclude_default}
\title{Default lengthy arguments}
\format{
A character string.
}
\usage{
chromToExclude_default
}
\description{
This contains a list of long constant values used as defaults in many function.
}
\examples{
REpatternFile_default  # Display the default value for REpatternFile.

chromToExclude_default
}
\keyword{datasets}
