% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRseek-package.R
\docType{package}
\name{CRISPRseek-package}
\alias{CRISPRseek}
\alias{CRISPRseek-package}
\title{CRISPRseek: Design of guide RNAs in CRISPR genome-editing systems}
\description{
The package encompasses functions to find potential guide RNAs for the CRISPR-based genome-editing systems including the Base Editors and the Prime Editors when supplied with target sequences as input. Users have the flexibility to filter resulting guide RNAs based on parameters such as the absence of restriction enzyme cut sites or the lack of paired guide RNAs. The package also facilitates genome-wide exploration for off-targets, offering features to score and rank off-targets, retrieve flanking sequences, and indicate whether the hits are located within exon regions. All detected guide RNAs are annotated with the cumulative scores of the top5 and topN off-targets together with the detailed information such as mismatch sites and restrictuion enzyme cut sites. The package also outputs INDELs and their frequencies for Cas9 targeted sites.
}
\keyword{internal}
