% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argsDefaults.R
\name{featureWeightMatrixFile_default}
\alias{featureWeightMatrixFile_default}
\title{featureWeightMatrixFile_default}
\usage{
featureWeightMatrixFile_default()
}
\description{
Default value for featureWeightMatrixFile, use featureWeightMatrixFile() to show its value.
}
