% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argsDefaults.R
\docType{data}
\name{gRNA.backbone_default}
\alias{gRNA.backbone_default}
\title{gRNA.backbone_default}
\format{
An object of class \code{character} of length 1.
}
\usage{
gRNA.backbone_default
}
\description{
gRNA.backbone_default
}
\keyword{datasets}
