% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argsDefaults.R
\name{mismatch.activity.file_default}
\alias{mismatch.activity.file_default}
\title{mismatch.activity.file_default}
\usage{
mismatch.activity.file_default()
}
\description{
Default value for mismatch.activity.file (csv), use mismatch.activity.file_default() to show its value
}
