% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyPlaceholder.R
\name{predIndelFreq}
\alias{predIndelFreq}
\title{Function definition place holders that are to be overwritten by reticulate
This is to suppress the R CMD check NOTE about "no visible global function""}
\usage{
predIndelFreq(thisSeq, weights)
}
\arguments{
\item{thisSeq}{param1 used in Lindel}

\item{weights}{param2 used in weights}
}
\description{
Function definition place holders that are to be overwritten by reticulate
This is to suppress the R CMD check NOTE about "no visible global function""
}
