% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argsDefaults.R
\name{REpatternFile_default}
\alias{REpatternFile_default}
\title{REpatternFile_default}
\usage{
REpatternFile_default()
}
\description{
Default value for REpatternFile, use REpatternFile_default() to show its value.
}
