% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argsDefaults.R
\docType{data}
\name{subPAM.activity_default}
\alias{subPAM.activity_default}
\title{subPAM.activity_default}
\format{
An object of class \code{hash} of length 16.
}
\usage{
subPAM.activity_default
}
\description{
subPAM.activity_default
}
\keyword{datasets}
