% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argsDefaults.R
\docType{data}
\name{weights_default}
\alias{weights_default}
\title{weights_default}
\format{
An object of class \code{numeric} of length 20.
}
\usage{
weights_default
}
\description{
weights_default
}
\keyword{datasets}
